Refazendo quase tudo porque o Joriver...

+++ backend/final/src/main/java/com/tipo_pgto/repository/TipoPgtoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoPgtoRepository extends JpaRepository<TipoPgto, Integer> {
+public interface TipoPgtoRepository extends JpaRepository<TipoPgto, Long> {
 }
